rpctest: Upstream create new rpctest package sync.
Upstream commit d04a6632cea1ed99ab33cb49ed587dc826aac787.

The merge commit contains necessary Decred-specific alterations and also
removes the code that is related to generating and submitting a new
block.

While the block generation functionality is useful, the specific method
used in the upstream code will not work in Decred because of the
additional state needed for voting and the fact the difficulty
adjustment period is much more frequent which prevents just using the
pow limit.

Future commits can remedy this situation by making use of block
templates provided by the daemon and allowing the rpc test code to
modify the template accordingly.