rpctest: Don't use Fatalf in goroutines.
This modifies the test code for joining mempools and blocks to send any
errors that occur over the sync channel and calling fatal in the main
goroutine instead of calling fatal in the goroutines to prevent any
potential teardown issues in the case of failure.

This also makes go vet happy.