dcrdtest: Remove need for *testing.T
This removes the need for passing a valid *testing.T object to the
harness initializing function New() by switching the logging statements
from using the test object to log it to using a standard slog logger.

The API is not yet changed to avoid having to perform a major version
bump in the package, but a deprecation note is added to the function's
documentation.

The purpose of this change is to allow this package to be more broadly
useful, not just in tests, but also in benchmarks and in generic code
not tied specifically to tests.