dcrdtest: Automatic TearDown on SetUp failures
This adds an automatic TearDown() call for an rpc harness that fails to
be setup, so that callers do not need to remember to do it themselves.

This improves usability of the package by ensuring goroutines do not
leak after either a failed SetUp() or a TearDown().

Unit tests are added that assert the correct behavior on both a
successfull setup and when certain specific setup calls fail.

This also fixes a goroutine leak in the internal memory wallet.