dcrdtest: Add option for keeping node dir
This adds a flag to the Harness to instruct it to keep the node dir
after TearDown and, unless that flag is set, removes the node dir.

This improves the reliability of the test harness by not leaving behind
extraneous, unneeded data by default.

When calling the TearDownInTest variant, if the test has already failed
then the node dir is kept to ease debugging test failures.